added fake swiping
